wibble: Compile with -Wall -Werror

+++ wibble/string.cpp
@@ -100,7 +100,7 @@ static const char invbase64(const T& idx)
 {
 	static const char data[] = {62,0,0,0,63,52,53,54,55,56,57,58,59,60,61,0,0,0,0,0,0,0,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,0,0,0,0,0,0,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51};
 	if (idx < 43) return 0;
-	if (idx > 43 + (sizeof(data)/sizeof(data[0]))) return 0;
+	if ((unsigned)idx > 43 + (sizeof(data)/sizeof(data[0]))) return 0;
 	return data[idx - 43];
 }
